+++ test/index.test.ts
@@ -15,9 +15,9 @@ const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
 describe("M3U8Downloader", () => {
   const m3u8Url = "http://127.0.0.1:3000/video.m3u8";
   const output = "/path/to/output.mp4";
-  const tempDir = path.join(os.tmpdir(), "m3u8-downloader");
-  if (!fs.existsSync(tempDir)) {
-    fs.mkdirSync(tempDir);
+  const segmentsDir = path.join(os.tmpdir(), "m3u8-downloader");
+  if (!fs.existsSync(segmentsDir)) {
+    fs.mkdirSync(segmentsDir);
   }
   serverStart();
   it("should express return", async () => {
@@ -47,11 +47,11 @@ describe("M3U8Downloader", () => {
 
   describe("download", () => {
     it("should download success", async ({ onTestFinished }) => {
-      const output = path.join(tempDir, "output.ts");
+      const output = path.join(segmentsDir, "output.ts");
 
       const downloader = new M3U8Downloader(m3u8Url, output, {
         convert2Mp4: false,
-        tempDir,
+        segmentsDir,
         // clean: false,
       }) as any;
 
@@ -70,11 +70,11 @@ describe("M3U8Downloader", () => {
     });
     it("should download segment failure", async () => {
       const m3u8Url = "http://127.0.0.1:3000/test.m3u8";
-      const output = path.join(tempDir, "output.ts");
+      const output = path.join(segmentsDir, "output.ts");
 
       const downloader = new M3U8Downloader(m3u8Url, output, {
         convert2Mp4: false,
-        tempDir,
+        segmentsDir,
       }) as any;
 
       await downloader.download();
@@ -82,18 +82,18 @@ describe("M3U8Downloader", () => {
 
       expect(downloader.status).toEqual("error");
       expect(fs.existsSync(output)).toBeFalsy();
-      expect(fs.existsSync(path.join(tempDir, "segment0.ts"))).toBeFalsy();
-      expect(fs.existsSync(path.join(tempDir, "segment1.ts"))).toBeFalsy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment0.ts"))).toBeFalsy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment1.ts"))).toBeFalsy();
     });
     it("should not merge when mergeSegments is false", async ({
       onTestFinished,
     }) => {
       const m3u8Url = "http://127.0.0.1:3000/video.m3u8";
-      const output = path.join(tempDir, "output.ts");
+      const output = path.join(segmentsDir, "output.ts");
 
       const downloader = new M3U8Downloader(m3u8Url, output, {
         mergeSegments: false,
-        tempDir,
+        segmentsDir,
       }) as any;
 
       await downloader.download();
@@ -101,27 +101,27 @@ describe("M3U8Downloader", () => {
 
       expect(downloader.status).toEqual("completed");
       expect(fs.existsSync(output)).toBeFalsy();
-      expect(fs.existsSync(path.join(tempDir, "segment0.ts"))).toBeTruthy();
-      expect(fs.existsSync(path.join(tempDir, "segment1.ts"))).toBeTruthy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment0.ts"))).toBeTruthy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment1.ts"))).toBeTruthy();
 
       onTestFinished(() => {
         // clean
-        if (fs.existsSync(path.join(tempDir, "segment0.ts"))) {
-          fs.unlinkSync(path.join(tempDir, "segment0.ts"));
+        if (fs.existsSync(path.join(segmentsDir, "segment0.ts"))) {
+          fs.unlinkSync(path.join(segmentsDir, "segment0.ts"));
         }
-        if (fs.existsSync(path.join(tempDir, "segment1.ts"))) {
-          fs.unlinkSync(path.join(tempDir, "segment1.ts"));
+        if (fs.existsSync(path.join(segmentsDir, "segment1.ts"))) {
+          fs.unlinkSync(path.join(segmentsDir, "segment1.ts"));
         }
       });
     });
   });
   describe("pause", () => {
     it("should pause the download", async ({ onTestFinished }) => {
-      const output = path.join(tempDir, "output.ts");
+      const output = path.join(segmentsDir, "output.ts");
 
       const downloader = new M3U8Downloader(m3u8Url, output, {
         convert2Mp4: false,
-        tempDir,
+        segmentsDir,
         headers: {
           delay: 1000,
         },
@@ -137,20 +137,20 @@ describe("M3U8Downloader", () => {
 
       onTestFinished(() => {
         // clean
-        if (fs.existsSync(path.join(tempDir, "segment0.ts"))) {
-          fs.unlinkSync(path.join(tempDir, "segment0.ts"));
+        if (fs.existsSync(path.join(segmentsDir, "segment0.ts"))) {
+          fs.unlinkSync(path.join(segmentsDir, "segment0.ts"));
         }
-        if (fs.existsSync(path.join(tempDir, "segment1.ts"))) {
-          fs.unlinkSync(path.join(tempDir, "segment1.ts"));
+        if (fs.existsSync(path.join(segmentsDir, "segment1.ts"))) {
+          fs.unlinkSync(path.join(segmentsDir, "segment1.ts"));
         }
       });
     });
     it("should pause the download and resume", async ({ onTestFinished }) => {
-      const output = path.join(tempDir, "output.ts");
+      const output = path.join(segmentsDir, "output.ts");
 
       const downloader = new M3U8Downloader(m3u8Url, output, {
         convert2Mp4: false,
-        tempDir,
+        segmentsDir,
         headers: {
           delay: 500,
           timeout: 5000,
@@ -183,27 +183,27 @@ describe("M3U8Downloader", () => {
       expect(hasPaused).toBeTruthy();
       expect(hasResumed).toBeTruthy();
 
-      expect(fs.existsSync(path.join(tempDir, "segment0.ts"))).toBeTruthy();
-      expect(fs.existsSync(path.join(tempDir, "segment1.ts"))).toBeTruthy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment0.ts"))).toBeTruthy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment1.ts"))).toBeTruthy();
 
       onTestFinished(() => {
         // clean
-        if (fs.existsSync(path.join(tempDir, "segment0.ts"))) {
-          fs.unlinkSync(path.join(tempDir, "segment0.ts"));
+        if (fs.existsSync(path.join(segmentsDir, "segment0.ts"))) {
+          fs.unlinkSync(path.join(segmentsDir, "segment0.ts"));
         }
-        if (fs.existsSync(path.join(tempDir, "segment1.ts"))) {
-          fs.unlinkSync(path.join(tempDir, "segment1.ts"));
+        if (fs.existsSync(path.join(segmentsDir, "segment1.ts"))) {
+          fs.unlinkSync(path.join(segmentsDir, "segment1.ts"));
         }
       });
     });
   });
   describe("cancel", () => {
     it("should cancel the download", async () => {
-      const output = path.join(tempDir, "output.ts");
+      const output = path.join(segmentsDir, "output.ts");
 
       const downloader = new M3U8Downloader(m3u8Url, output, {
         convert2Mp4: false,
-        tempDir,
+        segmentsDir,
         headers: {
           delay: 500,
         },
@@ -218,17 +218,17 @@ describe("M3U8Downloader", () => {
       await sleep(800);
       expect(downloader.downloadedFiles.length).toEqual(1);
       // has been auto clean
-      expect(fs.existsSync(path.join(tempDir, "segment0.ts"))).toBeFalsy();
-      expect(fs.existsSync(path.join(tempDir, "segment1.ts"))).toBeFalsy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment0.ts"))).toBeFalsy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment1.ts"))).toBeFalsy();
     });
     it("should not clean when cancel the download", async ({
       onTestFinished,
     }) => {
-      const output = path.join(tempDir, "output.ts");
+      const output = path.join(segmentsDir, "output.ts");
 
       const downloader = new M3U8Downloader(m3u8Url, output, {
         convert2Mp4: false,
-        tempDir,
+        segmentsDir,
         headers: {
           delay: 500,
         },
@@ -243,16 +243,16 @@ describe("M3U8Downloader", () => {
       await sleep(800);
       expect(downloader.downloadedFiles.length).toEqual(1);
       // has been auto clean
-      expect(fs.existsSync(path.join(tempDir, "segment0.ts"))).toBeTruthy();
-      expect(fs.existsSync(path.join(tempDir, "segment1.ts"))).toBeFalsy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment0.ts"))).toBeTruthy();
+      expect(fs.existsSync(path.join(segmentsDir, "segment1.ts"))).toBeFalsy();
 
       onTestFinished(() => {
         // clean
-        if (fs.existsSync(path.join(tempDir, "segment0.ts"))) {
-          fs.unlinkSync(path.join(tempDir, "segment0.ts"));
+        if (fs.existsSync(path.join(segmentsDir, "segment0.ts"))) {
+          fs.unlinkSync(path.join(segmentsDir, "segment0.ts"));
         }
-        if (fs.existsSync(path.join(tempDir, "segment1.ts"))) {
-          fs.unlinkSync(path.join(tempDir, "segment1.ts"));
+        if (fs.existsSync(path.join(segmentsDir, "segment1.ts"))) {
+          fs.unlinkSync(path.join(segmentsDir, "segment1.ts"));
         }
       });
     });
@@ -273,7 +273,7 @@ describe("M3U8Downloader", () => {
 
   it("should download the TS segments", async () => {
     const downloader = new M3U8Downloader(m3u8Url, output, {
-      tempDir,
+      segmentsDir,
     }) as any;
     downloader.status = "running";
     const tsUrls = [
@@ -287,24 +287,24 @@ describe("M3U8Downloader", () => {
     });
     await downloader.downloadTsSegments(tsUrls);
     expect(downloadSegment).toEqual(tsUrls.length);
-    expect(fs.existsSync(path.join(tempDir, "segment0.ts"))).toBeTruthy();
-    expect(fs.existsSync(path.join(tempDir, "segment1.ts"))).toBeTruthy();
+    expect(fs.existsSync(path.join(segmentsDir, "segment0.ts"))).toBeTruthy();
+    expect(fs.existsSync(path.join(segmentsDir, "segment1.ts"))).toBeTruthy();
 
     // clean
-    if (fs.existsSync(path.join(tempDir, "segment0.ts"))) {
-      fs.unlinkSync(path.join(tempDir, "segment0.ts"));
+    if (fs.existsSync(path.join(segmentsDir, "segment0.ts"))) {
+      fs.unlinkSync(path.join(segmentsDir, "segment0.ts"));
     }
-    if (fs.existsSync(path.join(tempDir, "segment1.ts"))) {
-      fs.unlinkSync(path.join(tempDir, "segment1.ts"));
+    if (fs.existsSync(path.join(segmentsDir, "segment1.ts"))) {
+      fs.unlinkSync(path.join(segmentsDir, "segment1.ts"));
     }
   });
 
   it("should merge the TS segments into a single file", async () => {
-    const tempDir = path.join(__dirname, "assets");
-    const output = path.join(tempDir, "output.ts");
+    const segmentsDir = path.join(__dirname, "assets");
+    const output = path.join(segmentsDir, "output.ts");
 
     const downloader = new M3U8Downloader(m3u8Url, output, {
-      tempDir,
+      segmentsDir,
     }) as any;
     downloader.status = "running";
 
@@ -323,7 +323,7 @@ describe("M3U8Downloader", () => {
 
   it("should convert the merged TS file to MP4", () => {
     const downloader = new M3U8Downloader(m3u8Url, output) as any;
-    const tsMediaPath = "/path/to/tempDir/output.ts";
+    const tsMediaPath = "/path/to/segmentsDir/output.ts";
 
     vi.spyOn(downloader, "convertToMp4").mockImplementation(() => {
       // Simulate converting to MP4