+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 4. add headers option
 5. set default userAgent as "Mozilla/5.0 (Windows NT 10.0; Win64; x64; rv:128.0) Gecko/20100101 Firefox/128.0"
 6. fix some bugs
-7. remove `segmentDownloaded` event
+
+## Breaking Change
+
+1. remove `segmentDownloaded` event
+2. change `tempDir` options to `segmentsDir`
 
 # 0.1.0